+++ server/src/request/NewBidRequest.java
@@ -1,5 +1,7 @@
 package request;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -7,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Getter @NoArgsConstructor @AllArgsConstructor
 public class NewBidRequest {
     private Integer auctionId;
-    private Long bidAmount;
+    private BigDecimal bidAmount;
 }
